Fix manage existing MissingDependencies error
When attempting to manage an existing volume a MissingDependencies error
is logged to c-vol.log. Some of the tasks defined in
ManageExistingTask#get_flow now require a 'volume' parameter not
the 'volume_ref' parameter.

To fix this issue, convert all 'volume_ref' parameter to 'volume'
parameter in managing existing volume task flow.

Co-Authored-By: wanghao <sxmatch1986@gmail.com>

Change-Id: I2b6ad5d86acba6c3a130275c9c3617ec70ae62e0
Closes-Bug: 1617442