VMAX driver - Retrieve volume from snapshot
The VMAX driver retrieves volume id from snapshot['volume_name'] in
create_snapshot and delete_snapshot and then gets volume from db using
the volume id.  If the volume_name is changed, volume id may not be in
it any more.  This is not a reliable way of getting volume id.
Also a driver should not access db directly. In this patch, the source
volume is retrieved from snapshot without parsing the volume name for
volume id and accessing db.

Change-Id: I99ebecb27d82d5a967369a04bf2c34158e6670ec
Closes-Bug: #1616133