VMAX driver - Won't delete pre-existing initiator groups.
If an initiator group has not been created by the VMAX
driver (i.e. pre-created on the array with a different
name to OS-<shorthostname>-<Protocol>-IG), it should not
be deleted by the driver if it has no more masking views
associated with it. This patch adds in a check to ensure
only initiator groups created by the VMAX driver are
deleted.

Change-Id: I5217ed4822590f00ae8c80d636e56905f320f6e8
Closes-Bug: #1618784