Allow getting qos_specs in db.volume_type_get
In our db.volume_type_get method we allow specifying expected_fields,
which are fields that are going to get joinloaded with the query, but we
only allow loading the projects and not the qos_specs.

This patch makes it possible to load the qos_specs related entries when
retrieving a volume type.

Change-Id: I9bf15359049266895db6a207621dc1281488b260