Allow attribute lazy loading in VolumeType OVO
VolumeType OVO class does not allow lazy loading of attributes, which is
usually not a problem if you are loading the volume type using
get_by_id, but is a problem if you load for example a list of volumes
and then go to their volume_type.extra_specs, as this will raise an
error because extra_specs cannot be retrieved.

This patch changes VolumeType OVO class to allow lazy loading all
optional fields.

Change-Id: Ief143a6c981cec4bdb21888776d610aa9d5dc9d8