Modifies override logic for key_manager
Makes the logic for overriding config options for the key_manager
more robust.

Before this patch, the override logic seemed to be called before the global
CONF object has been populated with values from the configuration file.
ConfKeyManager, the default for if no value had been specified, would be
used to override the value for api_class. Then when CONF was populated
with the actual values, the ConfKeyManager override value would still be
set.

This patch makes the logic a little bit more robust so that the value
is only overriden if explicitly passed into the function, not at the
global scope outside of the function.

SecurityImpact

Closes-Bug: 1621109
Change-Id: Id5f83f69fd3a877459fab924c005047e55f98c7b