Make sure get_backup_device result is an o.vo
In c-bak we're assuming we'll get a Snapshot or Volume o.vo from c-vol's
get_backup_device. This isn't true when we're running Mitaka's c-vol in
our environement. As we should be compatible with services in Mitaka
this commit adds a check to volume.rpcapi to make sure that if we'll
receive a dictionary, we're converting it to o.vo.

Change-Id: Ifab88181d78481e79bc913f90b3dd1be2cbf7400
Closes-Bug: 1621836