Fix creating typeless volumes with Mitaka's c-sch
In Mitaka as a workaround for creating volumes without volume type c-api
was sending an empty dictionary as request_spec['volume_type'] in
create_volume() RPC cast to scheduler. This changed in Newton and we're
handling this situation directly in scheduler, but when running Newton's
c-api with Mitaka's c-sch we should mimic older behavior. This patch
implements that.

Change-Id: I3dbb1934f3d8e68fef49f56155cbf59bc5fadc3a
Closes-Bug: 1619008
(cherry picked from commit 91bdf9de4c1e0726892384605704d87b318e7a69)