Fix RBD timeout
On the RBD driver the rados_connect_timeout is not working and the
default 300 seconds timeout is being applied.

This is a considerable problem if the connectivity to the cluster is
lost because cinder volume service will get stuck trying to retrieve the
stats and will appear as being down to the scheduler.

The reason for the timeout not being honored is because connect method
in librados library no longer accepts a timeout parameter, so the Python
bindings accept the timeout parameter but ignore it.

As can be read in the Python binding code [1]:
        # NOTE(sileht): timeout was supported by old python API, but
        # this is not something available in C API, so ignore for now
        # and remove it later

This patch fixes this by setting timeouts at the Rados class level.

[1]: https://github.com/ceph/ceph/blob/master/src/pybind/rados/rados.pyx#L793

Closes-Bug: #1621942
Change-Id: I3266e1331f5f4586b1e56ea78f6caaf8752d869f