Fix logic in retype check for encryption changes
The check was looking for not None to decide if it can
call the driver.retype(..). Turns out the diff has an
empty dictionary when it didn't find any encryption diff.

We now just look at the converted boolean value of whatever
was in there so the empty dict is interpreted the right way.

Change-Id: Ic1ae3c6596f2a390b8f8a41495d1a58ee58eacdf
Closes-Bug: #1622762