Dell SC: Disable connectivity failures for new servers
The Dell SC driver creates an object on the array called a server
definition for any new hosts it needs to expose a volume to. By default
these server definitions will have an option enabled called
AlertOnConnectivity that will cause alerts to be generated and emails to
be sent if configured.

Since OpenStack controls connecting and disconnecting iSCSI sessions as
needed, this results in false alarms being raised for servers. For
servers used in an OpenStack environment, this flag should be set to
false so alerts on not generated when a server logs out of a target
port.

No unit tests included since I couldn't find a way with existing tests
to not just test that mocking works.

Change-Id: I677aec1c8f24b068bbd95103f4d33d677df8875e
Closes-bug: #1623280