Add iscsi port to sf_svip arg if not provided
With things like VLAN support it may be necessary for
an admin to specify the sf_svip (IP) address.  The problem
is that when this is specified if the admin forgets to include
the port suffix to the SVIP address it may not work on certain
distros (inparticular RHEL variants).

This patch just adds a simple check when the endpoint is created
to make sure the port suffix is included, if it's not it
automatically adds the default 3260.

NOTE: This is not a backport, the Cheesecake code has an
inadvertent fix (albeit it could be improved) that didn't
merge in Mitaka.

Change-Id: Id2a99ff927d1fb9e40fab10beb0db5e4f302ae70
Closes-Bug: #1619941