Fix missing JSON response in API reference
Currently, the API reference is showing a blank sample response
for the v2 "List API Versions" call [1].  That appears to be because
the sample file it includes is "versions-resp.json" which exists, but
is 0 bytes (completely empty).  There is also a
"versions-response.json" file present in the samples directory, which
appears to contain a valid JSON reply.

This patch fixes the problem by removing the 0-byte sample file and
pointing to the correct one.

[1]
http://developer.openstack.org/api-ref/block-storage/v2/#list-api-versions

Change-Id: If2c1b486229681f68ba4ab0c13307de4599d9b72
Closes-Bug: #1623744