VMAX driver - Ensure VMAX volume matches cinder db volume
The VMAX cinder driver locates volumes on the backend by
using the volume's backend 'DeviceID'. It does not check
to ensure the volume retrieved matches the volume in the
cinder database. This can cause inconsistencies in cinder,
where a volume may be deleted off the backend, but remains
in the cinder database. This patch addresses the issue by
ensuring the 'ElementName' of the volume retrieved matches
the VMAX cinder driver element name ('OS-' + cinder UUID).

Change-Id: I8da37cd772d24fbb63a6fb12ccd782ffb4fef5c2
Closes-Bug: #1401297