Fix project assignment in VolumeType OVO
We have a problem with the assignment of the "project" field on the
VolumeType OVO class, because the DB queries return in some cases a list
of strings -for example in "volume_type_get" method- while on others
they return a VolumeTypeProjects SQLAlchemy model class.

The source of all this confussion is that some DB methods do not return
ORM instances but dictionaries with transformed data, which is something
that should never have been done in the first place.

This patch does not try to fix the underlying issue, just the incorrect
attribute assignment in the least intrussive way possible.

Change-Id: I0dd6b8ab2e726f69ceaaae886058e7a6181dded1
Closes-Bug: #1626704