VMAX driver - Duplicate initiator group name error.
As part of the validation of each of the components in a
masking view, the code checks if the initiators in the
initiator group associated with the masking view match
those from the connector. If the initiators don't match,
and there is no other initiator group associated with any
of the initiators from the connector, a new initiator
group with the correct initiators is created. However,
the current code does not first check if the ig to be
created has the same name as the ig from the masking
view. If so, the ig from the masking view must be first
deleted before a new ig with the same name can be created.

Change-Id: I51111c66b1954c82ed7df01d737d57c7bc57171e
Closes-Bug: #1579934