VMAX driver - PortGroup rollback error path scenario.
An errorMessage is being overwritten by
_get_port_group_name_from_mv if an existing masking view
component (the initiator group or storage group) could not
be verified, but the masking view and port group still
exist on the array. This would cause the errorMessage to
be set to None (as port group can successfully be
retrieved) and the code would continue on, causing errors
later. Even if there is an error in retrieving the port
group, _get_port_group_name_from_mv does not correctly
return the errorMessage variable to the calling function.

Change-Id: I20f5473ccfa59019fd9b638f85e9c89dccc54b20
Closes-Bug: #1578714