Disable API v1 by default
The v1 API has been deprecated by default since Juno. While we've
determined we can't get rid of it in the foreseeable future, we
shouldn't enable it by default. This may encourage any remaining
users to migrate.

This will also get rid of some warning messages in the logs on
startup warning of the deprecated API.

Change-Id: Ieea9a6c764eb1b55ad399a96e218d773ad4c59a5
Partial-bug: #1413102