HNAS: Rename hnas_svcX_volume_type
Hitachi NAS Platform (HNAS) drivers use the configuration option
hnas_svcX_volume_type (X is a number from 0 to 3) to get the name of
the pools configured for the drivers. The name containing
"volume_type" in the end confuses the users because they tend to
relate the option with the volume types created in cinder.
This patch deprecates the name hnas_svcX_volume_type when
configuring the drivers by cinder.conf in favor of
hnas_svcX_pool_name, so it can express more precisely the real
purpose of this option.
This modification applies only for those configuration using
the cinder.conf. The XML configuration options remain the same
since this kind of configuration is deprecated.

DocImpact

Change-Id: I4e2f145b1124f5b567585ec831a58ca6c8138aa5