Make Image Volume Cache cluster aware
Image Volume Cache mechanism was not cluster aware and therefore cached
images would not be shared among different services in the same cluster.

This patch addresses this issue and makes sure that they share by
cluster if there is one.

This patch does not address any concurrency issues that may currently
exist in the caching mechanism's code.

Implements: blueprint cinder-volume-active-active-support
Change-Id: I9be2b3c6dc571ce2e0e4ccf7557123a7858c1990