RBD: prevent creation of encrypted volumes
At the moment RBD volume encryption is not supported.  Libvirt needs
support for non-block encryption and nova will need a few changes as
well.  Efforts are underway to resolve this, but in the mean time it's
very confusing to users when we allow them to create encrypted volumes
using volume types as stated in the documentation.  Encryption of the
volume will eventually fail, but allowing this creation creates an
unreasonable expectation.

This patch prevents encrypted volume creation for RBD until support has
landed.  At that time, this patch should be reverted.

Change-Id: I0f071a4aea2d6454b71d83055e87d87b9efcc014