Dell SC: Add init_volume check and logging
Added a check to make sure the init_volume function uses
a physical server. Added a check and log message if the
volume was not able to be initialized.

Added a generic debug log message to find_volume.

Added a log message when the provider_id is no longer valid.

Change-Id: I385d3d8518ac5c60944f21d670f35e974eef28a0