Fix config generator issue with OVO decorators
Using an OVO decorator in cinder directly from cinder.objects makes the
configuration generator fail with an Attribute error.

The reason for this is that the OVOs have not been registered as they
should.

This patch fixes this by registering the OVOs first thing in the
cinder/opts.py file.

Change-Id: Id50330b9dc4b7a86a23b3c0fccef55aadfd09044
Closes-Bug: #1639199