Fix replication_status on InvalidReplicationTarget
When we cannot do a replication failover because the driver raises an
InvalidReplicationTarget exception we revert the replication status from
"failing-over" to its previous status.

Since we are not saving the previous status we were deciding what it was
based on the requested secondary_backend_id, which is wrong, as it
assumes we cannot failover from one secondary to another secondary.

This patch fixes this by changing the deciding factor from the
secondary_backend_id to the current backend_id in the service.

We had no failover_host manager tests, and this patch doesn't try to add
all the tests that we should have, it just adds the tests pertinent for
this patch to avoid regressions and test this specific functionality.
Missing tests should be added in another patch.

Closes-Bug: #1641715
Change-Id: I5a911278aef8060e14577099b0e03daf2039a783