Fix status after a VolumeDriverException
If a driver raised VolumeDriverException during failover_host then the
service would remain in "failing-over" replication status and we would
see an AttributeError exception in the logs.

That is because we were trying to set the "status" field in a service,
and that field doesn't exist.

This patch fixes this by setting the replication_status field to error
instead and by disabling the service.

We had no failover_host manager tests, and this patch doesn't try to add
all the tests that we should have, it just adds the tests pertinent for
this patch to avoid regressions and test this specific functionality.
Missing tests should be added in another patch.

Closes-Bug: #1641716
Change-Id: I374a8a187b93da5f946f0243c841ba0b54273401