Fix unit tests for LVM migration with os-brick 1.7.0
Commit a9f50dd3823ea2e9c568a1fdd970e268955dca9f in os-brick 1.7.0
introduced the usage of safe_decode() of the results of CLI commands.
This revealed that two unit tests in Cinder use a mocked "execute"
method of os-brick *without* specifying the return value. This lead
to an exception in the safe_decode() method as it isn't able to
decode MagicMocks:
    TypeError: <class 'mock.mock.MagicMock'> can't be decoded

This change adds return values to the mocked "execute" method which
can be decoded safely.

This change can not be cherry picked from the newton and ocata fixes
(dc84c1de76e51487c8b7a72ba623c12ee2747879) due to refactoring of the
test_volume unit test files.

Co-Authored-By: Markus Zoeller <mzoeller@de.ibm.com>

Change-Id: I9b0a6c5be41c471d1a7782f77927ffe3b9652043
Closes-Bug: 1638861