Make failover DB changes consistent
There are some inconsistencies among the different drivers when it comes
to setting the replication_status field in the DB on a failover.

To ensure that consistent behavior is provided the manager will oversee
all DB changes during failover and failback and make additional changes
when necessaries to maintain this consistency.

The drivers will no longer need to process non replicated volumes on
failover/failback and can simplify their logic as they will only receive
replicated volumes based on the replication_status field.

On failover:

- Non replicated volumes will have their status changed to error, have
  their current status saved to the previous_status field, and their
  replication_status changed to not_replicated.
- All non replicated snapshots will have their statuses changed to
  error.
- All replicated volumes that failed on the failover will get their
  status changed to error, their current status saved in
  previous_status, and their replication_status set to failover-error.
- All snapshots from volumes with failover errors will have their status
  set to error.
- All volumes successfully failed over will get their replication_status
  changed to failed-over.

On failback:

- All non replicated volumes and snapshots will have no model update.
- All replicated volumes that failed on the failover will get their
  status changed to error, their current status saved in
  previous_status, and their replication_status set to failover-error.
- All snapshots from volumes with failover errors will have their status
  set to error.
- All volumes successfully failed back will get their replication_status
  changed to enabled.

A more detailed explanation is provided in the updated replication
devref [3].

Since there were some drivers that where not setting the
replication_status on creation [1] and retype [2] we have added a
migration script to fix this inconsistency.

[1]: Fixed in change: I10a7931ed4a73dfd2b69f0db979bc32a71aedb11
[2]: Fixed in change: Id4df2b7ad55e9b5def91329f5437da9caa185c30
[3]: Change: I180b89c1ceaeea6d4da8e995e46181990d52825f

Closes-Bug: #1643886
Change-Id: Iefc409f2b19d8575a4ca1ec98a15276f5604eb8d