Check oslo.concurrency supports prlimits
Bug #1449062 was fixed and backported to Liberty by adding qemu limits settings
to oslo putils execute call. The problem is that min requirement for Liberty is
oslo.concurrency 2.30 and this feature requires oslo.concurrency 2.6.1.

oslo.concurrency isn't pinned but this could cause some issues if folks are on
the older version. We should just add a simple try/except block around this so
it doesn't break people that are running older versions of oslo.concurrency.

Change-Id: I9e3965980f28fc0760b81a6e2d60db33a794cca3
Closes-Bug: #1641992