Deprecate the block_device driver
The BlockDevice Driver continues to be behind,
and not tested very well; and adoption doesn't
appear to be that significant.

The only argument to having/keeping it in the past
has been that it was required for performance, however
most recent data analysis using the LIO driver and LVM
(standard reference driver) seems to indicate that there
in fact may not be a significant performance difference.

Additionally, if higher performance is desired, using a
striping across multiple disks with the LVM reference
drivers will provide higher performance.

Closes-Bug: #1642441
DocImpact

Change-Id: Ic5f8c49e9c707a812cd0dc1d99b1e46a06c60015