HPE3PAR: handle conflict in iscsi host create
One of the step in creating iSCSI VLUN is to create a host on array.
If there are simultaneous requests comming on 3par array for creating
same host, hpe3parclient creates host for first request and returns
HTTP 409 conflict error for others, with error code 73, message "host
WWN/iSCSI name already used by another host". This was not handled by
3PAR iscsi driver and the HTTP exception would propagate up to volume
manager where it would log it. This patch addresses issue by catching
this exception and reusing the same host.

Change-Id: I102409539c9a691c1816a342163dd049855f57da
Closes-Bug: #1642945