Move service and cluster creation in test to utils
In this patch we move service and cluster creation methods from the
tests to the test utils file so they can be easily reused by other tests
that need to create them.

This change is required by the patch that fixes the replication freeze
mechanism but wasn't included in that patch to facilitate the review by
splitting the 2 different concepts: moving these convenience methods to
test utils and fixing the freeze mechanism.

Related-Bug: #1616974
Change-Id: I7d8552f38e9495f72a5c1af61f4f57b3b4683157