SolidFire QoS scaled by volume size
This commit implements scaled QoS for the SolidFire driver. The
following QoS and Extra specs are now recognized: scaledIOPS, scaleMin,
scaleMax, scaleBurst. ScaledIOPS presence is a flag which causes the
other 3 to scale their respective QoS starting with the original
minIOPS, maxIOPS and burstIOPS by the size of the volume.

Moved the SolidFire tests to a separate directory to accomodate extra
files for data driven testing.

Change-Id: I83cd0a1b2f978351596c20ab2bfa6dc44776ed6d
Implements: blueprint solidfire-scaled-qos