Dell SC: Unable to locate LV primary after LV-AFO
The Dell SC driver does not always locate the LV primary
after an LV-AFO event. The details in the LiveVolume object
returned by the DSM after an LV-AFO event can vary depending
on state and SCOS versions.

In the case that where the LV has failed over, the secondary
is active and no primary volume is returned in the LV object,
the driver now returns the proper LV object.

Closes-Bug: #1645506
Change-Id: I4ebcf1e1030425855bd873e5a4c2305a3519ee2e