Route extend_volume calls to scheduler
When extending a volume in a THIN pool, Cinder has no checks of whether
the new allocated space is provisioned by the backend. This allow a
tenant, with enough quota space to extend the volume to a size much
higher than the provisioned space reported by the backend.

So, by routing the extend request through the scheduler, it will
assure that the host can provision the needed space.

Closes-bug: #1600295
Closes-bug: #1539112
Change-Id: I4a4ee8acac258d773b99e2767f1c75faa909813b