Add missing consistencygroup_id in volume
consistencygroup_id is missing from the volume object during
volume creation after it is provided by the user. This problem
was introduced by the migrating CG code. This patch fixed this
issue. It also made consistencygroup object available in the
volume object during volume creation and made consistencygroup
object available in the cgsnapshot object when cgsnapshot is
created/deleted.

Change-Id: I243f9306df883831165eaed450c76cb4c831787c
Closes-Bug: #1661393