Remove cgsnapshot_id before snapshot.save
There are a few more places where information is missing for
consistencygroup or cgsnapshot due to the CG migration changes.
This patch tries to fix them.

cgsnapshot_id should be removed from the snapshot object before
snapshot.save is called in the exception block because cgsnapshot
does not exist in the db.

cgsnapshot object is added to the snapshot object when snapshots
are passed to the driver in create/delete cgsnapshot.

Change-Id: I253a98b8a2058a00c2c577927c0b701c2f88c81b
Closes-Bug: #1662684
(cherry picked from commit 702868fd00e6b99d7ae5bda64d3410b2e5f9ddc5)