LVM: collect more debug data when VG creation fails
Our 3rd party CI fails sometimes because Cinder cannot create a
LVM volume group. The error in the c-vol log is:

    WARNING: lvmetad is running but disabled. Restart lvmetad before
    enabling it!\n  Volume group "stack-volumes-lvmdriver-1" has in
    sufficient free space (255 extents): 768 required.\n'

This failure cannot be pinned to one specific tempest test case.
In order to get more debug data when this case happens, this change
adds another error log entry to display the current state of the
volume groups on the system.

Change-Id: Ib75d20a1d603e620acf4766aef4f5876e98ab848
Related-Bug: 1186902