Remove run_tests.sh wrapper
The run_tests.sh mechanism for performing operations
is no longer the preferred approach. It is now better
to just run tox. There are also issues with enforcing
upper constraints that may cause test runs to fail.

To get rid of confusion by having two different ways
to run unit tests, flake8 checks, etc., it's probably
better to just remove the run_tests.sh script completely
and direct folks to use tox instead. This way we are all
consistent and there is less concern that changes will
be made to one and not the other.

Change-Id: I03f4b49cffd9f64e4eb3bdc27995d543be9efa43