HNAS: Change snapshot names
Currently, HNAS NFS driver creates snapshots named as
snapshot-<snapshot-id> in the backend. This is not very effective
because there is no precise way to relate the snapshot with the volume
that originated it using HNAS features.
This patch changes the snapshot related features to work with
snapshots named as <volume-name>.<snapshot-id> in order to make easier
to relate the snapshots with their parent volumes. The operations for
snapshots named the old way are still supported, but deprecated.

Co-Authored-By: Erlon Cruz <erlon.cruz@fit-tecnologia.org.br>
Change-Id: I188c7a3f67558e17ccc1c50e0f00dbb076395b18