Delete volume when rescheduling
If a reschedulable failure happens in the driver after the volume was
created on the backend, it will get orphaned and only way to remove it
will be through backend internals. This commit adds an additional
delete_volume call to the driver after the volume got rescheduled to
make sure no volumes will get orphaned.

Change-Id: Idd86a4842bdc6ecf0cabbeff0a9c9704e030302a
Closes-Bug: 1561579