DS8K driver: change the design of CG
When making consistency group snapshot, flashcopy in DS8K needs
to freeze the whole LSS which holds volumes. So it will affect
volumes not belonging to the CG if they are in the LSS used by
volumes belonging to CG. In order to dedicate LSS to CG, we need
to allocate volume not belonging to CG in other LSSs.

On the other hand, this change switches to generic volume group
from consistency group at the same time, but generic volume group
driver interfaces are added in another review, please refer to
https://review.openstack.org/#/c/445860.

DocImpact
Depends-On: Id9964f868c2dc14aaee37e4bcbba330fb4575492
Closes-Bug: #1655841
Change-Id: Ic6406a558c035ebb34177f87076adaec55b08f65