Deprecate the Linux SMBFS driver
We're marking the Linux SMBFS driver as unsupported, completely
removing it during the next release.

We've decided to deprecate it mostly because of its limitations
and lack of demand, unlike the Windows SMB driver which is largely
adopted in Hyper-V deployments.

Our CI will continue to send reports for this driver until it's
completely removed.

DocImpact

Change-Id: I1da58e36fc29a4dfee58454cf123543adbc6cac4