Fix volume manage
After moving the manage operation to support Active/Active the scheduler
no longer coud find the backend to place managed volumes.

The issue comes from the fact that the user can request managing a
volume just with the host, but if that backend is in a cluster the
volume in the DB needs to be created with it, so in the API we create
the volume using the service's host and cluster_name, but those are
missing the pool information.

The solution requires that the scheduler is capable of matching the
backends even if we are searching for a backend without the pool as well
as the scheduler updating the volume DB entry to include the pool.

Co-Authored-By: Erlon R. Cruz <erlon.cruz@fit-tecnologia.org.br>
Closes-bug: #1652811
Change-Id: I6110c0f93097f4e8c2b5fbc934924e3af1ec431b