VMware: Use full clone if extend needed
VMDK driver creates a new volume as a linked clone of the source
volume or snapshot if the extra-spec contains vmware:clone_type=
linked. The volume in vCenter cannot be extended if it's a linked
clone of another volume or snapshot. Therefore, using full clone
if volume needs to be extended after clone operation. Also
refactoring unit tests to reduce code duplication.

Closes-bug: #1660955
Change-Id: I9c111da431c8e3ef6de7e064e7399526cd02991e