Fix Qos for clone for Nimble Storage Driver
Currently the Qos is set only for the create volume.
The attribute needs to be overriden by the clone.
By default the clone inherits the attributes of the parent,
but if user needs to modify the Qos and Folder name of the child
(clone) they should be able to do it with this fix.

Change-Id: Ic30c1f386ee902333260fe23d57954ea021c24a4
Closes-Bug: #1659165