Dell EMC SC: Delete doesn't account for migration
The function delete_volume is unaware of migration. The volume object
sent down to delete_volume from retype with migration is a mix of the new
and old volumes. In addition to this the Dell EMC API function delete_volume
doesn't validate provider_id to ensure it is for the current back end.

Both of thse are fixed by this patch.

Change-Id: Ie927d2ee5f1f1a61d8309049ea7630f73f614cc5
Closes-Bug: #1659600