Zero out SolidFire capacity when unreachable
If for some reason connectivity to the cluster goes down, we just return the
cached capacity info.  Then the scheduler still sees the cluster as an
available resource in the pool.

This change detects connectivity issues during the get_cluster_stats call and
if the call fails, we report 0 available capacity to keep create calls from
being scheduled to the cluster.

Change-Id: I3f730e140c2b61fdd407c90b134916108312278a
Closes-Bug: #1398877