Fix attachments after attached migration
When migrating an attached volume, after the migration completion,
the volume gots stuck and can not be deleted or detached. This
is happening due an error in the volume completion implementation.
Now, before detaching the source volume, we first save the
attachments and then re-connect then in the destination volume.

Closes-bug: #1659914
Change-Id: Ie9a7815df99869b62209d99bfdcf51b250a8dba4
(cherry picked from commit a5ae53867bbe7f66bd640061a68a690c6d217177)