VMware: Use storage profile name in extra-spec
Currently we query vCenter for the storage profile name associated
with the volume. We should use the volume type extra-spec 'vmware:
storage_profile' as the source of truth because vCenter may associate
a default storage profile with the volume even if we do not specify
a storage profile (based on the value of vmware:storage_profile)
during volume creation.

Change-Id: Ib9f553a69262234e5329d79d20ef8a997758a89f
Closes-bug: #1660644