Fix host assignment when clustered
When creating a volume or a group we set the host that is actually
working on the resource on the volume service (this is not necessary,
but it's useful for operators), but the check and assignment methods in
manager's `_set_resource_host` are wrong, since we are comparing using =
a host with a pool and one without.

We are also incorrectly making the assignment because we are not adding
the pool to the host.

This patch fixes this by properly comparing the resource's host with our
own without the backend, and the adding the pool to our host before
doing the assignment.

Closes-Bug: #1660734
Change-Id: Id4c69f872135276e445d14fb36bc6bac0121a028