VMAX driver - ignore service level and workload in xml
When multi_pool_support is enabled in cinder.conf then
the driver should ignore any ServiceLevel entries in the
emc xml file. This was not happening because of an
incorrect check while fetching and populating the various
service level/workload combinations from VMAX. With the
fix, if any ServiceLevel or Workload tags are present in
the emc xml file, they will be ignored.

Change-Id: I45a8660aee2e4e4623aedc52c99ddeb51396980e
Closes-Bug: #1659831