Disable multiattach for all drivers
This patch disables setting multiattach flag to True in driver capabilities.
Since we reworked the driver attach/detach API, we need to ensure that the
issue of detaching a shared volume works properly on the Nova side.  A shared
flag might need to be required to return in initialize_connection to support
multiattach, so for now we will disable the capability.

Change-Id: I4f8229a4465d009e06de86da255f442020151113