3PAR: Remove un-necessary snapshot in online copy operation
While doing an online copy in the 3PAR, we create a temporary
snapshot of source volume to do the data copy operation.

Online copy is a feature in 3PAR whereby a copy initiated
on a source volume , and when the data copy is in progress,
the target volume of the copy is immediately available for
volume attachment and performing I/O is allowed on that volume
Resynching the source and the destination volume is
done automatically by the 3PAR internally.

Reasoning behind removing the temporary snapshot of the source
volume is that, internally 3PAR creates a temporary snapshot
and extra snapshot created by Cinder 3PAR Driver is un-necessary.

Also this temporary snapshot is not cleaned up on the
completion of online copy operation.

The presence of the temporary snapshot during the volume copy
operation is confirmed by running the "showtask" on the
asynchronous task id.

Change-Id: Ia2cb70e77211a372f73eaf577dbf57715c5a363f
Closes-Bug: 1661541