Fix action on services raises TypeError
Freeze, thaw and failover_host actions on a service raises TypeError
when non-existent host or cluster name is passed to the service
update API. This TypeError exception raised is converted into
HTTPBadRequest at wsgi layer.

Fixed this issue by formatting error message when a service is not
found for a given host or cluster. Caught the ServiceNotFound at API
layer and raised InvalidInput error so a user will see 400 error
only instead of 404.

Closes-Bug: #1663205
Change-Id: I54b9575f293f64a5f31246d0345c753b8011d95f