VMAX Driver - disable initiator check by default
For each existing masking view we check if the initiator group
matches with the initiators of the compute host. The is a very
expensive check and should be turned off by default. To turn
it back on we will put a initiator_check flag in the cinder.conf

Change-Id: Ia0677bafe9d586e9e65cd0d63924259f9a2e6ee8
Closes-Bug: #1663312
(cherry picked from commit 81aae7f71233c0297c503cbd79c3eaacc1a6b462)