Change nova_catalog_admin_info to default to publicURL
The only services in OpenStack that have dedicated admin REST APIs
that actually enforce different permissions at the API level are
keystone and swift. Assuming there is a dedicated admin endpoint is an
artifical construct.

Change the default for the admin_info to be the normal public url. For
sites that really want 2 different values here, they can have it, but
default to the simpler thing.

Change-Id: I92efc4639d37cb2736bae7e40707a853f1e72be6