VMware: Delete temporary snapshot after clone
VMDK driver creates the new volume as a linked clone of the
source volume if vmware:clone_type is set to 'linked' in the
volume type extra-spec. To create a linked clone, we should
create a temporary snapshot and then create the clone off the
snapshot point. After the clone, we can delete the temporary
snapshot, which we are not doing currently. Fixing it by
deleting the temporary snapshot in the finally block.

Closes-bug: #1664547
Change-Id: I3e198941f1e62af55133de2b3e2c46e898bcc2ed