VMware: Remove redundant sanity check
During create volume from a source volume or snapshot, we
validate the destination volume type extra-spec options.
This is unnecessary because the destination volume type
is always compatible with the source volume type which
is already validated. Removing the redundant checks to
validate the extra-spec options. Also refactoring related
test cases for readability and to re-use some of the
existing helper methods.

Change-Id: I3cfa2b7ba3d3c3b551e2d45e53d67e0a38f6685f