Add a README in tests dir linked to devref html
We have been adding docs and doing better at documenting
things under the docs section, but there are a number of
things still commonly missed by folks.  Testing info is
one of the more common, so let's just link the existing
testing doc to a README file in the tests directory to
make things more prominent.

Change-Id: I005309c50b7874da37ae5ab1180c2c8c6905c643