Remove reserving quotas in c-vol's retype method
Code removed by this commit was there to keep backward compatibility
between Liberty and Mitaka. As we're in Pike and we're only constrained
to keep compatibility with X-1 release, it's safe to get rid of it.
Right now we can assume that quota reservation will always be done in
c-api and passed down to c-vol.

Change-Id: I67b45662a52a55e88b920c2d0cbe9c20247a8d83