VMAX driver - volumes part of CG not managed correctly
Volumes created as part of a clone CG operation (either from the
source CG or a CG snapshot) were not managed properly. This was
because the provider_location was not being set as part of the
volume_model_update. With this fix, the provider_location is set
properly for the new target Volumes which are created as part of
the cloning of the CG. Additionally, a random UUID is generated
before creating the volume and is assigned as the name of the
volume on VMAX. This UUID is stored as a part of a key-value pair
in the volume.s admin metadata. This UUID is used later to uniquely
identify the volume on the VMAX while doing a volume operation.

Change-Id: I16ab7e25202c8ce861d8654097cf17fcf9bc8295
Closes-Bug: #1669827
(cherry picked from commit 5f5ddd9c40f3926dbe2360d4187c274ebeaf42b0)