Storwize: add CG capability to generic groups
This patch adds consistency group capability to generic volume
groups in Storwize driver. Re-use the CG implementations if it
is a CG type group. And if the group being created isn't a CG
then we bail out and the generic volume groups will take care of it.

Implements: blueprint storwize-generic-group
Change-Id: I64a7a29fd6620c3c8ea848c84bd564c9a632d307