Fix create_consistencygroup when using CG objects
When create_consistencygroup receives a ConsistencyGroup object
the parsing of the volume type ids produces an extra [] which
eventually causes the database query to fail.
This is relevant only to environments that are still using CG objects
instead of VG objects.

This code should be removed in Pike, and the goal of this change is to
fix the issue when backported to Ocata.

Change-Id: I2178e051200038a83abf55b5bd62137f155596df
Closes-Bug: 1668735
(cherry picked from commit 354b5d1ca454841f6b1c6a2a0620b2e9778fcfbe)