Fix host check in is_backend_frozen
When we fixed the freeze mechanism we introduced a DB function called
``is_backend_frozen`` that is used to check in the DB whether a resource
(volume, group, etc.) is frozen or not.

The issue is that this function is not taking into consideration that in
some cases host/cluster_name may be comming with the pool information,
so the check will fail.

This patch fixes this by removing the pool information from host and
cluster_name parameters before doing the check in the DB.

TrivialFix

Change-Id: Ie776adf9e746cf4cb7a2856d64d0b94423149b8d
Closes-Bug: #1616974