api-ref: Fix volume_id parameter for the path
When the volume_id is in the path for an API, it is
required. This fixes the volume_id_1 parameter for the
path argument by making it required and also renaming
it to be more clear of what it's used for. And finally,
this fixes the os-attach parameter to indicate volume_id
is in the path and is required, not optional.

Change-Id: I0e8868d35b5ea06ef809c896f943d92ad8e71e3e