Fix migrate api giving incorrect exception message
If the cluster is not configured and user has passed a non existent
host to the migrate api, it shows the below exception message:

"No available service named None"

The above exception message is not meaningful. While preparing an
error message to log, it fails to perform the 'or' operation
which results into the above error message.

For example:
>>> a = 'q'
>>> b = None
>>> c = 'I am %s' % b or a
>>> c
'I am None'
>>>
>>> c = 'I am %s' % (b or a)
>>> c
'I am q'

To fix this issue added the 'or' operation in bracket.

TrivialFix
Change-Id: I5d4673cc429bd9132c323bc63f111429de712050