NetApp cDOT: Support generic groups for block
Adding support for generic volume groups to NetApp cDOT's FC, and iSCSI
drivers.

CG methods are moved to the 7-mode block driver. Announcement was made
in a previous release for the removal of the NetApp 7-mode drivers.
Generic groups will only be implemented for the NetApp cDOT drivers.

Change-Id: I8290734817d171f6797c88d0a0d1a21f8b5789db
Implements: blueprint netapp-add-generic-group-support-cdot