HPE Lefthand: add CG capability to generic groups
This patch adds consistency group capability to generic volume group
in Lefthand driver. Re-use the CG implementations if it is a CG
type group. And if the group being created isn't a CG then we
bail out and the generic volume group will take care of it.

Implements: blueprint hpe-lefthand-generic-volume-group
Change-Id: Ib4b353331bad30e7296264542511b1d25a88cd0f