VNX: allow specify pool name for replication
Currently, VNX driver will select a pool of same name as the
source volume when creating LUN replication. An error will occur if not
satisfied.
This patch will allow setting a different pool name for replication
target instead of using the pool of same name on the remote system.

DocImpact

Closes-bug: 1677561
Change-Id: Ia86c9a5c815c19d0096524b88d992d84266688f9