Glance: attach volume encryption key id to image
This is required to be able to handle encrypted volumes
that have been uploaded to Glance.

Clone the volume's encryption key, and store the new encryption
key id in the image metadata in a property called
"cinder_encryption_key_id".  This allows the key to be retrieved
when creating a new volume from this image.

Related-Bug: #1485449
Related bp: improve-encrypted-volume

Change-Id: Ia1771817e6a06cc51c5357536915a2c5f9f6248e